+++ public/app/features/alerting/unified/RuleEditorGrafanaRecordingRules.test.tsx
@@ -2,11 +2,10 @@ import { UserEvent } from '@testing-library/user-event';
 import * as React from 'react';
 import { renderRuleEditor, ui } from 'test/helpers/alertingRuleEditor';
 import { clickSelectOption } from 'test/helpers/selectOptionInTest';
-import { screen } from 'test/test-utils';
+import { screen, testWithFeatureToggles } from 'test/test-utils';
 import { byRole } from 'testing-library-selector';
 
 import { FeatureToggles } from '@grafana/data';
-import { testWithFeatureToggles } from '@grafana/test-utils';
 import { contextSrv } from 'app/core/services/context_srv';
 import { setupMswServer } from 'app/features/alerting/unified/mockApi';
 import { PROMETHEUS_DATASOURCE_UID } from 'app/features/alerting/unified/mocks/server/constants';
