+++ public/app/features/alerting/unified/components/rules/deleted-rules/DeletedRules.test.tsx
@@ -1,7 +1,6 @@
 import { produce } from 'immer';
-import { render, screen } from 'test/test-utils';
+import { render, screen, testWithFeatureToggles } from 'test/test-utils';
 
-import { testWithFeatureToggles } from '@grafana/test-utils';
 import { AppNotificationList } from 'app/core/components/AppNotifications/AppNotificationList';
 import { DashboardSearchItemType } from 'app/features/search/types';
 import { AccessControlAction } from 'app/types/accessControl';
